requirements.txt
flask
requests
gunicorn
mysql-connector-python
pyyaml
